+++ tests/common/mod.rs
@@ -1,4 +1,5 @@
 use anyhow::{anyhow, Result as AnyhowResult};
+use mod_kit::mod_bam::CollapseMethod;
 use mod_kit::summarize::{summarize_modbam, ModSummary};
 use mod_kit::threshold_mod_caller::MultipleThresholdModCaller;
 use std::fs::File;
@@ -23,10 +24,11 @@ pub fn run_modkit(args: &[&str]) -> AnyhowResult<Output> {
     }
 }
 
-pub fn run_simple_summary(
+fn run_summary<'a>(
     bam_fp: &str,
     interval_size: u32,
-) -> AnyhowResult<ModSummary> {
+    collapse_method: Option<&CollapseMethod>,
+) -> AnyhowResult<ModSummary<'a>> {
     let threads = 1usize;
     let pool = rayon::ThreadPoolBuilder::new().num_threads(1).build()?;
     pool.install(|| {
@@ -41,9 +43,26 @@ pub fn run_simple_summary(
             0.1, // doesn't matter
             Some(MultipleThresholdModCaller::new_passthrough()),
             None,
+            collapse_method,
         )
     })
 }
+
+pub fn run_simple_summary(
+    bam_fp: &str,
+    interval_size: u32,
+) -> AnyhowResult<ModSummary> {
+    run_summary(bam_fp, interval_size, None)
+}
+
+pub fn run_simple_summary_with_collapse_method<'a>(
+    bam_fp: &str,
+    interval_size: u32,
+    collapse_method: &CollapseMethod,
+) -> AnyhowResult<ModSummary<'a>> {
+    run_summary(bam_fp, interval_size, Some(collapse_method))
+}
+
 pub fn check_against_expected_text_file(output_fp: &str, expected_fp: &str) {
     let test = {
         let mut fh = File::open(output_fp).unwrap();